* Moved header files into dedicated include directory. * Added “ORLP_” prefix to declspec macro (to prevent possible collisions with other libraries making use of the same name). * Added CMakeLists.txt and build.sh for easy packaging and inclusion into other projects. * Deleted .dll files inside repo.
