+++ .gitignore
@@ -1,0 +1,4 @@
+**/*.DS_Store
+.idea/
+*[Bb]uild*
+!build.sh
