+++ include/orlp-ed25519/fixedint.h
